requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pandas
openpyxl
celery[redis]
django-celery-beat
gunicorn
pytest
pytest-django